+++ scripts/linux/client-common.sh
@@ -2,8 +2,8 @@
 
 set -Eeuo pipefail
 
-export KIT_VERSION="v0.1.14"
-export DEFAULT_APP_VERSION="v0.1.8"
+export KIT_VERSION="v0.1.15"
+export DEFAULT_APP_VERSION="v0.1.9"
 export DOCKER_PLATFORM=""
 DOCKER_CMD=(docker)
 SUDO_CMD=()